Move text_to_image to utils
