+++ modules/rquiz.py
@@ -2,8 +2,9 @@ import os
 import pandas as pd
 import cutlet
 from dotenv import load_dotenv
-import discord
-from PIL import Image, ImageDraw, ImageFont
+
+# Utils
+import helper.utils as utils
 
 load_dotenv()
 katsu = cutlet.Cutlet()
@@ -85,7 +86,7 @@ class RQuiz:
     async def send_question(self, message):
         current_row = self.get_data().iloc[self.get_current()]
         # await message.channel.send("{number}. {kanji}".format(number=self.get_current() + 1, kanji=current_row["Kanji"]))
-        await text_to_image(message, current_row["Kanji"], self.get_current() + 1)
+        await utils.text_to_image(message, current_row["Kanji"], self.get_current() + 1)
 
     async def check_finished(self, message):
         if self.check_total():
@@ -103,16 +104,3 @@ async def start(message, data, data_key, index=0, length=10, max_try=3):
         n=10 if length == 0 else length).reset_index(drop=True), length, max_try)
     await message.channel.send("Japanese reading quiz started!")
     await data[data_key].send_question(message)
-
-
-async def text_to_image(message, text, number):
-    img = Image.new("RGB", (900, 250), color=(255, 255, 255))
-    fnt = ImageFont.truetype(os.path.abspath(
-        os.path.join((os.getcwd()), './fonts/meiryo.ttc')), 128)
-    d = ImageDraw.Draw(img)
-    d.text((30, 30), "{num}. {kanji}".format(
-        num=number, kanji=text), font=fnt, fill=(0, 0, 0))
-    filedir = os.path.abspath(os.path.join(
-        os.getcwd(), "./modules/cache/rquiz-temp.png"))
-    img.save(filedir)
-    await message.channel.send(file=discord.File(filedir))
